pyproject.toml
[tool.poetry]
name = "dttto"
version = "0.0.1"
description = "Discord Ticket Tool Transcript Overview"
authors = ["bastirlnd"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-login = "^0.6.3"
flask-bcrypt = "^1.0.1"
flask-wtf = "^1.2.1"
flask-migrate = "^4.0.7"
flask-testing = "^0.8.1"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
email-validator = "^2.1.1"
pre-commit = "^3.7.0"
black = "^24.4.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
